+++ docs/web/docs/reference/Blueprints.md
@@ -1,0 +1,16 @@
+---
+sidebar_position: 2
+---
+
+
+# Blueprints
+
+## Static React Task
+
+## Static Task
+
+## Remote Procedure
+
+## Parlai Chat
+
+## Mock
