+++ docs/web/docs/reference/overview.md
@@ -1,3 +1,7 @@
+---
+sidebar_position: 1
+---
+
 import Link from '@docusaurus/Link';
 
 # Overview
