+++ pyproject.toml
@@ -67,6 +67,7 @@ click_default_group = "^1.2"
 tabulate = "^0.8.7"
 rich = "^12.5.1"
 rich-click = "^1.5.1"
+mdutils = "^1.4.0"
 
 # [tool.poetry.group.metrics.dependencies]
 prometheus-client = "^0.13.1"
